worker/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
